Restrict total label count to satisfy new limit
Occasionally, pull requests to WPT modify many files across many
directories. That causes this project to assign a large number of
labels, although in practice, modifications of this type are
infrastructural in nature and do not typically concern specification
compliance.

As of 2018-07-26, GitHub.com restricts the total number of pull request
labels to 100. Ensure that no more than 100 labels are ever assigned to
a pull request by replacing over-long lists with the labels `infra` and
`very-large`.