Implement epochal update hook
This introduces a new `/update_epoch` endpoint that updates the `epochs/[epoch]` branch ref in `web-platform-tests/wpt` iff it has changed.

The request format expected from the wpt.fyi revision announcer is verified, but no authentication is required; the revision announcer API (rather than the incoming request body) is trusted to deliver an accurate epochal hash.

This is towards #41 and web-platform-tests/wpt.fyi#551.