Add support for WebKit-exported changes (#120)
This commit adds automation for changes exported from WebKit. It adds the appropriate labels
for WebKit PRs, marks them as auto-approved, and monitors their status in WebKit (aka reviewed,
committed).

Note that the actual commenting + labeling is currently disabled and will be enabled in a follow-up
commit. Additionally, future work will add the ability to auto-land an exported WebKit PR when the
downstream version is committed.

* Added webkit metadata library
* Added bugs-webkit library for WebKit's Bugzilla
* Added pullRequestPuller() in the main index.js. This function pulls
  the "Webkit export" open PR and updates the metadata info verifiying
  the PR against the Bugzilla
* Added support for get and set the bugsWebkitToken in the Google Cloud
  datastore
* getMedatada() extended to include WebKit related items:
  * For WebKit's Bugzilla flags: in_queue and reviewed
  * For WebKit's Bugzilla see_also link reference
* Tests:
  * Added tests for PR filtered by label
  * Added tests for retrieval and formatting of metadata for a WebKit
    related pull request
  * Added tests for PR with wrong WebKit bugzilla reference