Enable (dry-run guarded) code to comment/label WebKit exports
Currently the code is hard-coded to run in dry-run mode, until we've
proved out that it should do what we expect.

Also reduces the frequency from every minute to every 5 minutes, since
we will be doing the same thing over-and-over (since we're in dry-run
mode and so nothing will actually change).