Ship the webkit-export support
This removes the hard-coded dryRun flag for labelling and approving
WebKit-exported PRs, and uses the normal dry-run support instead.
Effectively, this launches the feature (since prod doesn't run in
dry-run mode!).

The 5-minute interval is left in place; I plan to remove it when we're
confident that the bot is doing the right thing.